Bash-Release:   4.0 Patch-ID:       bash40-028
Bug-Reported-by:        martin f krafft <madduck@debian.org>
Bug-Reference-ID:       <4A4E39E7.5080807@debian.org>
Bug-Reference-URL:      http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=519165
                        http://lists.gnu.org/archive/html/bug-bash/2009-07/msg00011.html

Bug-Description:

bash-4.0 reverted to the historical shell behavior of raising an error
when $@ or $* was expanded after `set -u' had been executed and there
were no positional parameters.  The Posix working group has since
clarified the standard's position on the issue, and $@ and $* are now the
only variables, parameters, or special parameters that do not raise an
error when unset if set -u is enabled.